Bigger size for paymentvalue to allow large wei values

+++ database_schema.sql
@@ -55,7 +55,7 @@ CREATE TABLE payment_log (
   tx_identifier  CHARACTER VARYING(255) PRIMARY KEY,
   creation_date  TIMESTAMP WITHOUT TIME ZONE NOT NULL,
   currency       CHARACTER VARYING(255) NOT NULL,
-  paymentvalue   BIGINT NOT NULL,
+  paymentvalue   NUMERIC(30, 0) NOT NULL,
   fx_rate        NUMERIC(19, 2),
   usd            NUMERIC(19, 2),
   email          CHARACTER VARYING(255)
